Fix pip-installability
