+++ requirements.txt
@@ -5,7 +5,6 @@ six
 mock
 nose2
 zmesh
-enum34
 
 # cloud-volume
 brotli
